Split generic geo test into multiple
