+++ test/celeritas/AllGeoTypedTestBase.hh
@@ -11,20 +11,12 @@
 #include <gtest/gtest.h>
 
 #include "celeritas_config.h"
-#include "orange/OrangeData.hh"
-#include "orange/OrangeParams.hh"
-#include "orange/OrangeTrackView.hh"
-
-#include "GenericGeoTestBase.hh"
+#include "celeritas/OrangeTestBase.hh"
 #if CELERITAS_USE_VECGEOM
-#    include "celeritas/ext/VecgeomData.hh"
-#    include "celeritas/ext/VecgeomParams.hh"
-#    include "celeritas/ext/VecgeomTrackView.hh"
+#    include "celeritas/ext/VecgeomTestBase.hh"
 #endif
 #if CELERITAS_USE_GEANT4
-#    include "celeritas/ext/GeantGeoData.hh"
-#    include "celeritas/ext/GeantGeoParams.hh"
-#    include "celeritas/ext/GeantGeoTrackView.hh"
+#    include "celeritas/ext/GeantGeoTestBase.hh"
 #endif
 
 namespace celeritas
@@ -39,7 +31,7 @@ namespace test
  *
  * To use this class to test all available geometry types, add
  * \code
- *   ${_needs_geo} LINK_LIBRARIES ${_geo_libs}
+ *   ${_needs_geo} LINK_LIBRARIES ${_all_geo_libs}
  * \endcode
  *
  * to the \c celeritas_add_test argument in CMakeLists.txt, and instantiate all
@@ -68,10 +60,7 @@ class AllGeoTypedTestBase : public GenericGeoTestBase<HP>
   public:
     using SPConstGeo = typename GenericGeoTestBase<HP>::SPConstGeo;
 
-    static std::string geo_name()
-    {
-        return testdetail::GenericGeoTraits<HP>::name;
-    }
+    static std::string geo_name() { return GenericGeoTraits<HP>::name; }
 
     SPConstGeo build_geometry() override
     {
@@ -82,12 +71,6 @@ class AllGeoTypedTestBase : public GenericGeoTestBase<HP>
 //---------------------------------------------------------------------------//
 // TYPE ALIASES
 //---------------------------------------------------------------------------//
-
-using GenericVecgeomTestBase = GenericGeoTestBase<VecgeomParams>;
-using GenericOrangeTestBase = GenericGeoTestBase<OrangeParams>;
-using GenericGeantGeoTestBase = GenericGeoTestBase<GeantGeoParams>;
-
-using GenericCoreGeoTestBase = GenericGeoTestBase<GeoParams>;
 
 using AllGeoTestingTypes = ::testing::Types<
 #if CELERITAS_USE_VECGEOM
@@ -104,7 +87,7 @@ struct AllGeoTestingTypeNames
     template<class U>
     static std::string GetName(int)
     {
-        return testdetail::GenericGeoTraits<U>::name;
+        return GenericGeoTraits<U>::name;
     }
 };
