+++ test/celeritas/CheckedGeoTrackView.t.hh
@@ -1,18 +1,20 @@
 //----------------------------------*-C++-*----------------------------------//
-// Copyright 2023-2024 UT-Battelle, LLC, and other Celeritas developers.
+// Copyright 2024 UT-Battelle, LLC, and other Celeritas developers.
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file celeritas/geo/CheckedGeoTrackView.hh
+//! \file celeritas/CheckedGeoTrackView.t.hh
+//! \brief Template definitions for CheckedGeoTrackView.
 //---------------------------------------------------------------------------//
 #pragma once
+
+#include "CheckedGeoTrackView.hh"
 
 #include <iomanip>
 
 #include "corecel/Types.hh"
 #include "corecel/cont/ArrayIO.hh"
 #include "corecel/io/Logger.hh"
-#include "orange/Types.hh"
 
 namespace celeritas
 {
@@ -20,100 +22,30 @@ namespace test
 {
 //---------------------------------------------------------------------------//
 /*!
- * Check validity of safety and volume crossings while navigating on CPU.
- *
- * Also count the number of calls to "find distance" and "find safety".
+ * Initialize the state.
  */
 template<class GTV>
-class CheckedGeoTrackView : public GTV
+CheckedGeoTrackView<GTV>&
+CheckedGeoTrackView<GTV>::operator=(GeoTrackInitializer const& init)
 {
-  public:
-    //!@{
-    //! \name Type aliases
-    using GeoTrackViewT = GTV;
-    using Initializer_t = typename GTV::Initializer_t;
-    using DetailedInitializer = typename GTV::DetailedInitializer;
-    //!@}
-
-  public:
-    //! Forward construction arguments to the original track view
-    template<class... Args>
-    CheckedGeoTrackView(Args&&... args) : GTV(std::forward<Args>(args)...)
-    {
-    }
-
-    // Initialize the state
-    inline CELER_FUNCTION CheckedGeoTrackView&
-    operator=(GeoTrackInitializer const& init)
-    {
-        GTV::operator=(init);
-        CELER_ENSURE(!this->is_outside());
-        return *this;
-    }
-    // Initialize the state from a parent state and new direction
-    inline CELER_FUNCTION CheckedGeoTrackView&
-    operator=(DetailedInitializer const& init)
-    {
-        GTV::operator=(init);
-        CELER_ENSURE(!this->is_outside());
-        return *this;
-    }
-
-    // Check volume consistency this far from the boundary
-    static constexpr real_type safety_tol()
-    {
-        return 1e-4 * units::centimeter;
-    }
-
-    // Calculate or return the safety up to an infinite distance
-    inline real_type find_safety();
-
-    // Calculate or return the safety up to the given distance
-    inline real_type find_safety(real_type max_safety);
-
-    // Change the direction
-    inline void set_dir(Real3 const&);
-
-    // Find the distance to the next boundary
-    inline Propagation find_next_step();
-
-    // Find the distance to the next boundary
-    inline Propagation find_next_step(real_type max_distance);
-
-    // Move a linear step fraction
-    inline void move_internal(real_type);
-
-    // Move within the safety distance to a specific point
-    inline void move_internal(Real3 const& pos);
-
-    // Move to the boundary in preparation for crossing it
-    inline void move_to_boundary();
-
-    // Cross from one side of the current surface to the other
-    inline void cross_boundary();
-
-    //! Number of calls fo find_next_step
-    size_type intersect_count() const { return num_intersect_; }
-    //! Number of calls fo find_safety
-    size_type safety_count() const { return num_safety_; }
-    //! Reset the stepscounter
-    void reset_count() { num_intersect_ = num_safety_ = 0; }
-
-  private:
-    bool checked_internal_{false};
-    size_type num_intersect_{0};
-    size_type num_safety_{0};
-};
+    GTV::operator=(init);
+    CELER_ENSURE(!this->is_outside());
+    return *this;
+}
 
 //---------------------------------------------------------------------------//
-// DEDUCTION GUIDES
-//---------------------------------------------------------------------------//
-
+/*!
+ * Initialize the state from a parent state and new direction.
+ */
 template<class GTV>
-CheckedGeoTrackView(GTV&&) -> CheckedGeoTrackView<GTV>;
+CheckedGeoTrackView<GTV>&
+CheckedGeoTrackView<GTV>::operator=(DetailedInitializer const& init)
+{
+    GTV::operator=(init);
+    CELER_ENSURE(!this->is_outside());
+    return *this;
+}
 
-//---------------------------------------------------------------------------//
-// INLINE DEFINITIONS
 //---------------------------------------------------------------------------//
 /*!
  * Calculate the safety distance.
