+++ test/celeritas/OrangeTestBase.hh
@@ -1,0 +1,36 @@
+//----------------------------------*-C++-*----------------------------------//
+// Copyright 2024 UT-Battelle, LLC, and other Celeritas developers.
+// See the top-level COPYRIGHT file for details.
+// SPDX-License-Identifier: (Apache-2.0 OR MIT)
+//---------------------------------------------------------------------------//
+//! \file celeritas/GenericOrangeTestBase.hh
+//---------------------------------------------------------------------------//
+#pragma once
+
+#include "orange/OrangeData.hh"
+#include "orange/OrangeParams.hh"
+#include "orange/OrangeTrackView.hh"
+
+#include "GenericGeoTestBase.hh"
+
+namespace celeritas
+{
+namespace test
+{
+//---------------------------------------------------------------------------//
+template<>
+struct GenericGeoTraits<OrangeParams>
+{
+    template<MemSpace M>
+    using StateStore = CollectionStateStore<OrangeStateData, M>;
+
+    using TrackView = OrangeTrackView;
+    static inline char const* ext = ".org.json";
+    static inline char const* name = "ORANGE";
+};
+
+using OrangeTestBase = GenericGeoTestBase<OrangeParams>;
+
+//---------------------------------------------------------------------------//
+}  // namespace test
+}  // namespace celeritas
